jam-spam-ml/requirements.txt
octokitpy
python-dotenv
requests
multi-rake
tensorflow
